--- a/flint-core/src/main/scala/org/opensearch/flint/core/storage/OpenSearchScrollReader.java
+++ b/flint-core/src/main/scala/org/opensearch/flint/core/storage/OpenSearchScrollReader.java
@@ -19,6 +19,8 @@ import org.opensearch.search.builder.SearchSourceBuilder;
 
 import java.io.IOException;
 import java.util.Optional;
+import java.util.function.BiFunction;
+import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -35,8 +37,23 @@ public class OpenSearchScrollReader extends OpenSearchReader {
 
   private String scrollId = null;
 
-  public OpenSearchScrollReader(IRestHighLevelClient client, String indexName, SearchSourceBuilder searchSourceBuilder, FlintOptions options) {
-    super(client, new SearchRequest().indices(indexName).source(searchSourceBuilder.size(options.getScrollSize())));
+  public OpenSearchScrollReader(
+      IRestHighLevelClient client,
+      String indexName,
+      SearchSourceBuilder searchSourceBuilder,
+      FlintOptions options) {
+    this(client, indexName, searchSourceBuilder, options, null);
+  }
+
+  public OpenSearchScrollReader(
+      IRestHighLevelClient client,
+      String indexName,
+      SearchSourceBuilder searchSourceBuilder,
+      FlintOptions options,
+      String preference) {
+    super(client,
+        applyPreference(preference).apply(new SearchRequest().indices(indexName)
+            .source(searchSourceBuilder.size(options.getScrollSize()))));
     this.options = options;
     this.scrollDuration = TimeValue.timeValueMinutes(options.getScrollDuration());
   }
@@ -52,8 +69,8 @@ public class OpenSearchScrollReader extends OpenSearchReader {
       return Optional.of(response);
     } else {
       try {
-        return Optional
-        .of(client.scroll(new SearchScrollRequest().scroll(scrollDuration).scrollId(scrollId),
+        return Optional.of(client.scroll(new SearchScrollRequest().scroll(scrollDuration)
+                .scrollId(scrollId),
             RequestOptions.DEFAULT));
       } catch (OpenSearchStatusException e) {
         LOG.log(Level.WARNING, "scroll context not exist", e);
@@ -75,8 +92,10 @@ public class OpenSearchScrollReader extends OpenSearchReader {
       }
     } catch (OpenSearchStatusException e) {
       // OpenSearch throw exception if scroll already closed. https://github.com/opensearch-project/OpenSearch/issues/11121
-      LOG.log(Level.WARNING, "close scroll exception, it is a known bug https://github" +
-          ".com/opensearch-project/OpenSearch/issues/11121.", e);
+      LOG.log(
+          Level.WARNING,
+          "close scroll exception, it is a known bug https://github" + ".com/opensearch-project/OpenSearch/issues/11121.",
+          e);
     } finally {
       scrollId = null;
     }
@@ -87,5 +106,13 @@ public class OpenSearchScrollReader extends OpenSearchReader {
    */
   public String getScrollId() {
     return scrollId;
+  }
+
+  static private Function<SearchRequest, SearchRequest> applyPreference(String preference) {
+    if (Strings.isNullOrEmpty(preference)) {
+      return searchRequest -> searchRequest;
+    } else {
+      return searchRequest -> searchRequest.preference(preference);
+    }
   }
 }