--- a/flint-spark-integration/src/main/scala/org/apache/spark/opensearch/table/ShardInfo.scala
+++ b/flint-spark-integration/src/main/scala/org/apache/spark/opensearch/table/ShardInfo.scala
@@ -1,0 +1,16 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+package org.apache.spark.opensearch.table
+
+/**
+ * Represents information about a shard in OpenSearch.
+ *
+ * @param indexName
+ *   The name of the index.
+ * @param id
+ *   The ID of the shard.
+ */
+case class ShardInfo(indexName: String, id: Int)