--- a/flint-spark-integration/src/main/scala/org/apache/spark/sql/flint/FlintScanBuilder.scala
+++ b/flint-spark-integration/src/main/scala/org/apache/spark/sql/flint/FlintScanBuilder.scala
@@ -6,13 +6,14 @@
 package org.apache.spark.sql.flint
 
 import org.apache.spark.internal.Logging
+import org.apache.spark.opensearch.table.OpenSearchTable
 import org.apache.spark.sql.connector.expressions.filter.Predicate
 import org.apache.spark.sql.connector.read.{Scan, ScanBuilder, SupportsPushDownV2Filters}
 import org.apache.spark.sql.flint.config.FlintSparkConf
 import org.apache.spark.sql.flint.storage.FlintQueryCompiler
 import org.apache.spark.sql.types.StructType
 
-case class FlintScanBuilder(tableName: String, schema: StructType, options: FlintSparkConf)
+case class FlintScanBuilder(table: OpenSearchTable, schema: StructType, options: FlintSparkConf)
     extends ScanBuilder
     with SupportsPushDownV2Filters
     with Logging {
@@ -20,7 +21,7 @@ case class FlintScanBuilder(tableName: String, schema: StructType, options: Flin
   private var pushedPredicate = Array.empty[Predicate]
 
   override def build(): Scan = {
-    FlintScan(tableName, schema, options, pushedPredicate)
+    FlintScan(table, schema, options, pushedPredicate)
   }
 
   override def pushPredicates(predicates: Array[Predicate]): Array[Predicate] = {